+++ api-admin/controllers/notificationlog/index.js
@@ -1,0 +1,10 @@
+
+import express from 'express';
+import list from './list';
+
+let router = express.Router();
+
+router.route('/notificationLog')
+    .get(list);
+
+module.exports = router;