+++ CHANGELOG.md
@@ -5,6 +5,7 @@ NOWnews-api change logs
 ## [Unreleased]
 ### Added
 - 網站推播時自帶 utm @esbb48
+- 增加推播紀錄端點 [GET] `/notificationLog` @esbb48
 
 ### Changed
 - 使用 pm2 config 控管 @esbb48